requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai==0.28.1
langchain-pinecone 
langchain-openai 
langchain
pandas 
numpy
python-dotenv
streamlit